Added Double/Float formatters
I added new formatter objets for Doubles and Floats. These
equivalent formatters are actually in Play's master branch,
so I'm thinkng that they could be removed in the future.

I also created "decimal" and "doubleDecimal" so you can
actually use them.

+++ site/app/models/Report.scala
@@ -6,7 +6,7 @@ import play.api.db.DB
 import play.api.Play.current
 
 case class Report(
-	id: Option[Int],
+	id: Option[Long],
 	latitude: Double,
 	longitude: Double,
 	address: String,
